Fix the testThreadSafety test.
WARNING: API incompatible change.

Having mutable enums is just a really bad idea.
I don’t really see any other way to fix the problem. The getInput() seems to only be used internally, but if this is considered a part of the public API… well, not sure what to do.