Introduce a cleaner lower level API with the Console interface.
It provides a signal api, access to the underlying terminal capabilities and stty settings.
It also brings support for virtual consoles when using remote connections for example.

It’s not much leveraged in the remaining of the code.  Things like cursor movements, etc.. should leverage this information.
The AeshInputStream hacks for transforming windows arrow keys can be just removed, and it should also be noted that the AeshInputStream does not correctly handle the encoding of the input stream and assumes the default charset, which is always not the case, especially on windows.

The only external dependency on those 3 new packages (api, impl, utils) is the LoggerUtil class.
This means that this can easily be extracted as the very-low component.