Put description in its own line for cli printHelp
The old behaviour of printing the help text is as follows:

```
Usage: command description

Options:
  --help
```

This is confusing because in the 'Usage' line, we're expecting the
cli-options to appear, and one might think that 'description' is a CLI
sub-command instead of the description itself (especially if the
description is short).

Instead, this commit puts the description in a new line to avoid this
confusion.

```
Usage: command
description

Options:
  --help
```